Refactored the code. most of the main class text is now in the uiInvoker class.
